+++ app/controllers/transfers/projects_controller.rb
@@ -60,7 +60,8 @@ class Transfers::ProjectsController < ApplicationController
       :handover_note_body,
       :two_requires_improvement,
       :inadequate_ofsted,
-      :financial_safeguarding_governance_issues
+      :financial_safeguarding_governance_issues,
+      :outgoing_trust_to_close
     )
   end
 end