+++ app/forms/transfer/create_project_form.rb
@@ -5,6 +5,7 @@ class Transfer::CreateProjectForm < CreateProjectForm
   attribute :two_requires_improvement, :boolean
   attribute :inadequate_ofsted, :boolean
   attribute :financial_safeguarding_governance_issues, :boolean
+  attribute :outgoing_trust_to_close, :boolean
 
   validates :outgoing_trust_ukprn, presence: true, ukprn: true
   validates :provisional_transfer_date, presence: true
@@ -12,6 +13,7 @@ class Transfer::CreateProjectForm < CreateProjectForm
   validates :two_requires_improvement, inclusion: {in: [true, false], message: I18n.t("errors.transfer_project.attributes.two_requires_improvement.inclusion")}
   validates :inadequate_ofsted, inclusion: {in: [true, false], message: I18n.t("errors.transfer_project.attributes.inadequate_ofsted.inclusion")}
   validates :financial_safeguarding_governance_issues, inclusion: {in: [true, false], message: I18n.t("errors.transfer_project.attributes.financial_safeguarding_governance_issues.inclusion")}
+  validates :outgoing_trust_to_close, inclusion: {in: [true, false], message: I18n.t("errors.transfer_project.attributes.outgoing_trust_to_close.inclusion")}
 
   validate :urn_unique_for_in_progress_transfers, if: -> { urn.present? }
 
@@ -52,7 +54,8 @@ class Transfer::CreateProjectForm < CreateProjectForm
       @note = Note.create(body: handover_note_body, project: @project, user: user, task_identifier: :handover) if handover_note_body
       @project.tasks_data.update!(
         inadequate_ofsted: inadequate_ofsted,
-        financial_safeguarding_governance_issues: financial_safeguarding_governance_issues
+        financial_safeguarding_governance_issues: financial_safeguarding_governance_issues,
+        outgoing_trust_to_close: outgoing_trust_to_close
       )
     end