+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - transfer projects now collect if they are the result of financial,
   safeguarding or governance issues and who this information on the about
   project tab
+- transfer projects now collect if the outgoing trust is expected to close once
+  the transfer is completed and this is shown on the about the project tab
 
 ## [Release-46][release-46]